Remove undefined behaviour
Shift left by 32 on a 32 bit type (std::size_t is 32 bit on 32bit
systems) is UB. The fix is not perfect, because it will not report the
correct result on 32bit systems, but I can live with that.

+++ include/osmium/io/bzip2_compression.hpp
@@ -230,7 +230,7 @@ namespace osmium {
                     if (bzerror != BZ_OK) {
                         throw bzip2_error{"bzip2 error: write close failed", bzerror};
                     }
-                    m_file_size = static_cast<std::size_t>(nbytes_out_hi32) << 32U | nbytes_out_lo32;
+                    m_file_size = static_cast<std::size_t>(static_cast<uint64_t>(nbytes_out_hi32) << 32U | nbytes_out_lo32);
                 }
             }
